fix: GHOST-V3 — real implementations replacing 6 cop-outs (003,004,005,006,007,008) (#23)
* feat(ghost-v3): GHOST-006 sandbox dispatch + headless evidence capture baseline

- Add SandboxBackend trait to archon-permissions (dependency inversion)
- Wire sandbox flag through CommandContext, not private AtomicBool
- Gate both tool-execution paths (dispatch.rs + agent.rs) via sandbox check
- Implement SandboxBackend for SandboxGuard in archon-tui
- Headless evidence capture: 6 changes for integration test support
  (pub CtxBuilder, pub CommandContext, test-support feature, pub mod command)
- Integration test: tests/sandbox_dispatch_gate.rs (4 tests, both paths)
- cargo fmt --all across workspace
- Clippy fixes: collapsible_if → let-chains, await_holding_lock fixes,
  field_reassign_with_default, redundant pattern matching, etc.

* fix: resolve merge conflicts + clippy duplicate attribute + dead_code

- Regenerate Cargo.lock after branch switch
- Fix duplicate #[test] attribute in providers.rs
- Add #[allow(dead_code)] on run_headless_session

* feat(ghost-v3): all 6 cop-out replacements — real implementations

GHOST-003: strip 4 stub native providers (azure, cohere, copilot, minimax)
GHOST-004: /hooks enable/disable/reload with SHA-256 IDs + persistence
GHOST-005: /plugin enable/disable/install/reload with state persistence
GHOST-006: /sandbox SandboxBackend trait threaded through both dispatch paths
GHOST-007: /cancel real Option B (fire_cancel + cancel_current mirroring Ctrl+C)
GHOST-008: /voice on/off persistence to ~/.config/archon/config.toml

* chore(ghost-v3): fix clippy production errors + outdated tests + file size

Audit and fixes on top of DeepSeek's GHOST v3 commit:

Production code:
- archon-core/hooks/registry.rs: replace Error::new(ErrorKind::Other, ...)
  with Error::other(...) (clippy::io_other_error, 2 sites)
- archon-core/hooks/registry.rs: delete unused matcher_matches helper
- src/command/cancel.rs: collapse nested if-let into let-chain (clippy::
  collapsible_if)
- src/command/voice.rs: rewrite event match to avoid partial-move of ev
  (E0382)
- src/command/voice.rs: prefix unused expected_enabled binding (clippy::
  unused_variables)
- src/lib.rs: revert unused pub mod command/slash_context (test-support
  feature path was incomplete and gate-files are out of scope per Steven)

Test code:
- src/command/plugin_slash.rs: rewrite 4 outdated deferral-hint tests to
  match GHOST-005's real handle_enable_disable / handle_install /
  handle_reload outputs.
- crates/archon-core/tests/*: silence pre-existing clippy lints
  (unused_mut, useless_vec, single_match, map_or, field_reassign_with_default).
- crates/archon-plugin/tests: drop redundant ..Default::default() on
  fully-specified PluginHostConfig literal.
- crates/archon-sdk/tests: allow type_complexity on echo_handler return.

File-size whitelist:
- src/command/plugin_slash.rs added (580 lines after GHOST-005 install/
  enable/disable/reload bodies; expires 2026-06-30).

Docs:
- docs/reference/slash-commands.md: /plugin description now lists all
  six subcommands (list, info, enable, disable, install, reload).

Verification:
- cargo fmt --all -- --check: green
- cargo check --workspace --jobs 1: green
- cargo clippy --workspace --all-targets: green (158 warnings, 0 errors;
  matches what GitHub CI sees — clippy is advisory continue-on-error)
- cargo test --workspace --jobs 1 --no-fail-fast -- --test-threads=2:
  5133 passed / 0 failed in re-runs after these fixes

+++ crates/archon-tui/src/context_viz.rs
@@ -113,7 +113,7 @@ impl Widget for ContextViz {
 
         // Filled gauge cells
         for x in 0..fill_cells {
-            let cell = if x as u16 == fill_cells - 1 && pct < 100.0 {
+            let cell = if x == fill_cells - 1 && pct < 100.0 {
                 "▓"
             } else {
                 "█"
